Filter tests from report if its skipped because of ocs version mismatch
