+++ ocs_ci/framework/pytest_customization/marks.py
@@ -105,3 +105,6 @@ gather_metrics_on_fail = pytest.mark.gather_metrics_on_fail
 
 # here is the place to implement some plugins hooks which will process marks
 # if some operation needs to be done for some specific marked tests.
+
+# Marker for skipping tests based on OCS version
+skipif_ocs_version = pytest.mark.skipif_ocs_version
